+++ lib/src/chart/base/axis_chart/axis_chart_data.dart
@@ -1124,13 +1124,14 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
   /// size, ... of the text.
   /// Drawing text will retrieve through [labelResolver],
   /// you can override it with your custom data.
-  /// /// [show] determines showing label or not.
+  /// [show] determines showing label or not.
   HorizontalLineLabel({
     EdgeInsets? padding,
     super.style,
     Alignment? alignment,
     super.show = false,
     String Function(HorizontalLine)? labelResolver,
+    super.angle = 0.0,
   })  : labelResolver =
             labelResolver ?? HorizontalLineLabel.defaultLineLabelResolver,
         super(
@@ -1158,6 +1159,7 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
       alignment: Alignment.lerp(a.alignment, b.alignment, t),
       labelResolver: b.labelResolver,
       show: b.show,
+      angle: lerpDouble(a.angle, b.angle, t)!,
     );
   }
 
@@ -1169,6 +1171,7 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
         padding,
         style,
         alignment,
+        angle,
       ];
 }
 
@@ -1186,6 +1189,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
     Alignment? alignment,
     bool? show,
     String Function(VerticalLine)? labelResolver,
+    super.angle = 0.0,
   })  : labelResolver =
             labelResolver ?? VerticalLineLabel.defaultLineLabelResolver,
         super(
@@ -1220,6 +1224,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
       alignment: Alignment.lerp(a.alignment, b.alignment, t),
       labelResolver: b.labelResolver,
       show: b.show,
+      angle: lerpDouble(a.angle, b.angle, t)!,
     );
   }
 
@@ -1231,6 +1236,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
         padding,
         style,
         alignment,
+        angle,
       ];
 }