+++ lib/src/chart/base/axis_chart/axis_chart_painter.dart
@@ -315,6 +315,7 @@ abstract class AxisChartPainter<D extends AxisChartData>
                 to.dy + padding.top,
               ),
             ),
+            label.angle,
           );
         }
       }
@@ -408,6 +409,7 @@ abstract class AxisChartPainter<D extends AxisChartData>
                 to.dy - padding.bottom,
               ),
             ),
+            label.angle,
           );
         }
       }